passage de rad à degré

+++ Projet/CamReg_correction/pi_regulator.h
@@ -1,9 +1,10 @@
 #ifndef PI_REGULATOR_H
 #define PI_REGULATOR_H
 
-int direction_haut(float);
+int direction_haut(int);
 //float get_angle(void);
 //start the PI regulator thread
 void pi_regulator_start(void);
 
 #endif /* PI_REGULATOR_H */
+
